Move query parser in its own mod
